pyproject.toml
[project]
name = "slm-fine-tuning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.65.0",
    "torch>=2.0",
    "transformers>=4.30",
    "scikit-learn>=1.3",
    "numpy>=1.24",
    "pandas>=2.0",
    "mlflow>=2.10",
    "extract-msg>=0.48",
    "streamlit>=1.30",
    "google-api-python-client>=2.100",
    "google-auth-oauthlib>=1.1",
    "google-auth-httplib2>=0.2",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.2.1",
    "ruff>=0.15.4",
    "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
